moved Toolbox::logInFile tests
